refactor: ScoreCalculator 제거 및 HandCards로 책임 이전
